Fix kappa calculation test
